+++ odoo/local-src/report_page_endnote/models/report_page_endnote.py
@@ -4,7 +4,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import re
-from openerp import models, fields, api
+from odoo import models, fields, api
 
 
 class ReportPageEndNote(models.Model):