+++ odoo/local-src/sale_discount_program/models/program_condition.py
@@ -2,9 +2,9 @@
 # © 2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import _, api, fields, models
-from openerp.addons import decimal_precision as dp
-from openerp.models import MAGIC_COLUMNS
+from odoo import _, api, fields, models
+from odoo.addons import decimal_precision as dp
+from odoo.models import MAGIC_COLUMNS
 
 
 class DiscountProgramCondition(models.Model):