+++ odoo/local-src/sale_discount_program/models/res_config.py
@@ -2,7 +2,7 @@
 # © 2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class SaleConfig(models.TransientModel):