+++ odoo/local-src/sale_discount_program/tests/test_discount_program.py
@@ -3,7 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 from datetime import timedelta, date
 
-from openerp.tests.common import TransactionCase, post_install, at_install
+from odoo.tests.common import TransactionCase, post_install, at_install
 
 
 class TestDiscountProgram(TransactionCase):