+++ odoo/local-src/specific_account/__manifest__.py
@@ -1,0 +1,20 @@
+# -*- coding: utf-8 -*-
+# © 2016 Julien Coux (Camptocamp)
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+{
+    'name': 'Depil Tech Account Customization',
+    'version': '10.0.1.0.0',
+    'author': 'Camptocamp',
+    'license': 'AGPL-3',
+    'category': 'Specific',
+    'website': 'http://www.camptocamp.com',
+    'images': [],
+    'depends': [
+        'base',
+        'account',
+    ],
+    'data': [],
+    'test': [],
+    'installable': True,
+    'auto_install': False,
+}