+++ odoo/local-src/specific_account/models/account_payment.py
@@ -2,7 +2,9 @@
 # © 2016 Julien Coux (Camptocamp)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp import models, api
+# TODO: Fix this class
+# TODO: Probably linked with 'account_tax_exigible' TODO
+from odoo import models, api
 
 
 class AccountPayment(models.Model):
@@ -19,14 +21,11 @@ class AccountMoveLine(models.Model):
     _inherit = 'account.move.line'
 
     @api.model
-    def create(self, vals, apply_taxes=True):
+    def create(self, vals):
         new_vals = vals.copy()
         ctx = self.env.context or {}
         if ctx.get('copy_partner_on_move') and not new_vals.get('partner_id'):
             new_vals['partner_id'] = ctx['copy_partner_on_move'].id
         if ctx.get('copy_ref_on_move') and not vals.get('ref'):
             new_vals['ref'] = ctx['copy_ref_on_move']
-        return super(AccountMoveLine, self).create(
-            new_vals,
-            apply_taxes=apply_taxes
-        )
+        return super(AccountMoveLine, self).create(new_vals)