+++ odoo/local-src/specific_discount_program/models/res_config.py
@@ -2,9 +2,9 @@
 # © 2016 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-from openerp.addons import decimal_precision as dp
+from odoo.addons import decimal_precision as dp
 
-from openerp import api, fields, models
+from odoo import api, fields, models
 
 
 class SaleConfig(models.TransientModel):