+++ odoo/local-src/specific_payment_mode/models/sale_order.py
@@ -2,11 +2,11 @@
 # © 2016 Julien Coux (Camptocamp)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-import openerp.addons.decimal_precision as dp
+import odoo.addons.decimal_precision as dp
 
-from openerp.exceptions import ValidationError
+from odoo.exceptions import ValidationError
 
-from openerp import models, fields, api, _
+from odoo import models, fields, api, _
 
 
 class SaleOrder(models.Model):