+++ odoo/songs/install/company.py
@@ -3,19 +3,19 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
 import anthem
-from base64 import b64encode
-from pkg_resources import Requirement, resource_string
 
 
 @anthem.log
 def setup_admin_companies(ctx):
+    """ setup_admin_companies """
     companies = ctx.env['res.company'].search([])
     admin = ctx.env['res.users'].search([('login', '=', 'admin')])
     admin.company_ids = companies.ids
 
 
 @anthem.log
-def remove_company_for_partners_of_coampnies(ctx):
+def remove_company_for_partners_of_companies(ctx):
+    """ remove_company_for_partners_of_companies """
     companies = ctx.env['res.company'].search([])
     for company in companies:
         partner = company.partner_id
@@ -25,6 +25,7 @@ def remove_company_for_partners_of_coampnies(ctx):
 
 @anthem.log
 def setup_company_report_footer(ctx):
+    """ setup_company_report_footer """
     companies = ctx.env['res.company'].search([])
     companies.write({
         'custom_footer': True,
@@ -37,19 +38,7 @@ def setup_company_report_footer(ctx):
 
 
 @anthem.log
-def setup_company_logo(ctx):
-    """ Setup company logo """
-    company = ctx.env.ref('base.main_company')
-
-    # load logo on company
-    req = Requirement.parse('depiltech-odoo')
-    logo_content = resource_string(req, 'data/images/logo.png')
-    b64_logo = b64encode(logo_content)
-    company.logo = b64_logo
-
-
-@anthem.log
 def main(ctx):
     setup_admin_companies(ctx)
-    remove_company_for_partners_of_coampnies(ctx)
+    remove_company_for_partners_of_companies(ctx)
     setup_company_report_footer(ctx)