+++ travis/publish.sh
@@ -1,40 +1,44 @@
 #!/bin/bash -e
 
+local_dir="$(dirname "$0")"
+
 function deploy {
-    local stack_name=$1
+    local tag=$1
 
-    wget -O - https://releases.rancher.com/compose/v0.12.1/rancher-compose-linux-amd64-v0.12.1.tar.gz |\
-        tar -x -z -C ${HOME} && mv ${HOME}/rancher-compose*/rancher-compose ${HOME}/ || exit $?
+    echo "Pushing image to docker hub ${DOCKER_HUB_REPO}:${tag}"
+    docker tag ${GENERATED_IMAGE} ${DOCKER_HUB_REPO}:${tag}
+    docker push ${DOCKER_HUB_REPO}:${tag}
 
-    RANCHER_COMPOSE="${HOME}/rancher-compose"
-    TEMPLATE_DIR="${TRAVIS_BUILD_DIR}/rancher/${stack_name}"
+    echo "Creating a minion for ${tag} on ${TRAVIS_BRANCH}"
+    $local_dir/minion-client.py \
+      ${tag} \
+      ${RANCHER_MINION_SERVER} \
+      ${minion_server_token} \
+      $local_dir/minion-files/docker-compose.yml \
+      $local_dir/minion-files/rancher-compose.yml \
+      $local_dir/minion-files/rancher.list
 
-    source <(echo $rancher_env_password | gpg --passphrase-fd 0 --decrypt --no-tty $TEMPLATE_DIR/rancher.env.gpg)
-    source $TEMPLATE_DIR/rancher.public.env
-
-    (cd "${TEMPLATE_DIR}" && \
-     ${RANCHER_COMPOSE} -p "${stack_name}" rm odoo db --force && \
-     sleep 30 && \
-     ${RANCHER_COMPOSE} -p "${stack_name}" up --pull --recreate --force-recreate --confirm-upgrade -d)
 }
 
 if [ "$TRAVIS_PULL_REQUEST" == "false" ]; then
   docker login --username="$DOCKER_USERNAME" --password="$DOCKER_PASSWORD"
+  docker_tag=r-$TRAVIS_BRANCH-$TRAVIS_COMMIT
 
   if [ "$TRAVIS_BRANCH" == "master" ]; then
-    echo "Deploying image to docker hub for master (latest)"
-    docker tag depiltech_odoo camptocamp/depiltech_odoo:latest
-    docker push "camptocamp/depiltech_odoo:latest"
-    # echo "Building test server"
-    # deploy $RANCHER_STACK_NAME
+    echo "Pushing image to docker hub ${DOCKER_HUB_REPO}:latest"
+    docker tag ${GENERATED_IMAGE} ${DOCKER_HUB_REPO}:latest
+    docker push "${DOCKER_HUB_REPO}:latest"
+
+    deploy ${docker_tag}
+
   elif [ ! -z "$TRAVIS_TAG" ]; then
-    echo "Deploying image to docker hub for tag ${TRAVIS_TAG}"
-    docker tag depiltech_odoo camptocamp/depiltech_odoo:${TRAVIS_TAG}
-    docker push "camptocamp/depiltech_odoo:${TRAVIS_TAG}"
+    echo "Pushing image to docker hub ${DOCKER_HUB_REPO}:${TRAVIS_TAG}"
+    docker tag ${GENERATED_IMAGE} ${DOCKER_HUB_REPO}:${TRAVIS_TAG}
+    docker push "${DOCKER_HUB_REPO}:${TRAVIS_TAG}"
+
   elif [ ! -z "$TRAVIS_BRANCH" ]; then
-    echo "Deploying image to docker hub for branch ${TRAVIS_BRANCH}"
-    docker tag depiltech_odoo camptocamp/depiltech_odoo:${TRAVIS_BRANCH}
-    docker push "camptocamp/depiltech_odoo:${TRAVIS_BRANCH}"
+    deploy ${docker_tag}
+
   else
     echo "Not deploying image"
   fi